+++ packages/pont-engine/src/constants/index.ts
@@ -1,0 +1,2 @@
+export const API_LOCK = 'api-lock.json';
+export const CONFIG_FILE = 'pont-config.json';
