+++ packages/pont-engine/src/main/Mocks.ts
@@ -1,0 +1,10 @@
+export class Mocks {
+  enable = false;
+  port = 8080;
+  basePath = '';
+  wrapper = `{
+      "code": 0,
+      "data": {response},
+      "message": ""
+    }`;
+}
