--- packages/pont-engine/src/utilTypes.ts
@@ -1,1 +1,0 @@
-export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
